LayoutTests/platform/ios-simulator-wk1/http/tests/dom/window-open-about-webkit-org-and-access-document-expected.txt

PASS newWindow.document is undefined.
